Fully working output fron neural net. Need to implement simulation now...
-Nick Buchert

+++ FurbiesFighters/src/neat/helpers/Helpers.java
@@ -8,22 +8,22 @@ import java.util.Map;
 import edu.furbiesfighters.skills.Skills;
 
 public class Helpers {
-	public static List<Float> generateOneHot(Enum choice, int size){
-		List<Float> oneHot = new ArrayList<Float>();
+	public static List<Double> generateOneHot(Enum choice, int size){
+		List<Double> oneHot = new ArrayList<Double>();
 		try {
 			// System.out.println("TEST " + (String[])choice.getClass().getMethod("values", null).invoke(choice));
 			
 			int selectionChoice = choice.ordinal(); 
 			for(int i = 0; i < size; i++) {
 				if(selectionChoice == i) 
-					oneHot.add(1.0f);
+					oneHot.add(1.0);
 				else
-					oneHot.add(0.0f);
+					oneHot.add(0.0);
 			}
 		} catch(Exception e) {
 			int i = 0;
 			while(i < size) {
-				oneHot.add(0.0f);
+				oneHot.add(0.0);
 				i++;
 			}
 		}
@@ -31,13 +31,13 @@ public class Helpers {
 		return oneHot;
 	}
 	
-	public static List<Float> boundedRechargeTime(Map<Skills, Integer> recharging){
-		List<Float> boundedTimes = new ArrayList<Float>();
+	public static List<Double> boundedRechargeTime(Map<Skills, Integer> recharging){
+		List<Double> boundedTimes = new ArrayList<Double>();
 		Skills[] skillOrder = new Skills[] {Skills.PAPER_CUT, Skills.REVERSAL_OF_FORTUNE, Skills.ROCK_THROW,
 		                  Skills.SCISSOR_POKE, Skills.SHOOT_THE_MOON};
 		
 		for(Skills skill : skillOrder) {
-			boundedTimes.add((float)recharging.get(skill) / 5);
+			boundedTimes.add((double)recharging.get(skill) / 5.0);
 		}
 		
 		return boundedTimes;
